# public/openapi.yaml
openapi: 3.0.0
info:
  title: Beantown Baby Diaper Bank API
  version: 1.0.0
  description: OpenAPI docs aligned with implemented Next.js API routes.
servers:
  - url: /api
    description: Internal API
paths:
...
  /total-diapers:
    get:
      summary: Get aggregate diaper total
      responses:
        '200':
          description: Total diapers
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalDiapers:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...